requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
django-safedelete~=1.4.0
flake8~=7.1.1
psycopg2~=2.9.10